@@ -1,21 +1,15 @@
-// Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
 import { getFirestore } from "firebase/firestore";
 
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const firebaseConfig = {
-  apiKey: "AIzaSyBWDb9qqpKj2NYs3K9Cw6qjlToEFl9RRAA",
-  authDomain: "boredify-ce61b.firebaseapp.com",
-  projectId: "boredify-ce61b",
-  storageBucket: "boredify-ce61b.appspot.com",
-  messagingSenderId: "180921146737",
-  appId: "1:180921146737:web:fd4bf14f6c9eb93573aad0",
-  measurementId: "G-NJWKHVZH4G",
-};
+const app = initializeApp({
+  apiKey: import.meta.env.VITE_API_KEY,
+  authDomain: import.meta.env.VITE_AUTH_DOMAIN,
+  projectId: import.meta.env.VITE_PROJECT_ID,
+  storageBucket: import.meta.env.VITE_STORAGE_BUCKET,
+  messagingSenderId: import.meta.env.VITE_MESSAGING_SENDER_ID,
+  appId: import.meta.env.VITE_APP_ID,
+});
 
 // Initialize Firebase
-const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
-
 export default db;
