@@ -1,6 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import logo from "../assets/images/logo.png";
+
+import logo from "~/assets/images/logo.png";
 
 function Navbar() {
   return (
